First pass at splitting into modules.
